+++ src/main/java/org/practicum/yandex/controller/dto/PostDto.java
@@ -3,13 +3,12 @@ package org.practicum.yandex.controller.dto;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Data
 @Builder
 public class PostDto {
-    private BigInteger id;
+    private Long id;
     private String title;
     private String text;
     private List<String> tags;
